LayoutTests/imported/mozilla/css-animations/test_animation-reverse-expected.txt


FAIL reverse() from idle state starts playing the animation assert_equals: expected (object) null but got (number) 0
